+++ src/main/java/com/kingpixel/cobblenotify/CobbleNotify.java
@@ -1,14 +1,21 @@
 package com.kingpixel.cobblenotify;
 
+import com.google.common.util.concurrent.ThreadFactoryBuilder;
 import com.kingpixel.cobblenotify.command.CommandTree;
 import com.kingpixel.cobblenotify.config.Config;
 import com.kingpixel.cobblenotify.config.Lang;
 import com.kingpixel.cobblenotify.config.Notifications;
+import com.kingpixel.cobblenotify.database.DataBaseClient;
+import com.kingpixel.cobblenotify.database.DataBaseFactory;
 import com.kingpixel.cobblenotify.events.Events;
 import dev.architectury.event.events.common.CommandRegistrationEvent;
 import dev.architectury.event.events.common.LifecycleEvent;
 import net.fabricmc.api.ModInitializer;
 import net.minecraft.server.MinecraftServer;
+
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
 /**
  * @author Carlos Varas Alonso - 28/04/2024 23:50
@@ -20,6 +27,12 @@ public class CobbleNotify implements ModInitializer {
   public static MinecraftServer server;
   public static Lang lang = new Lang();
   public static Config config = new Config();
+  public static DataBaseClient databaseClient;
+  private static final ExecutorService EXECUTOR_SERVICE = Executors.newFixedThreadPool(2, new ThreadFactoryBuilder()
+    .setDaemon(true)
+    .setNameFormat("CobbleNotify-Executor-%d")
+    .build()
+  );
 
   @Override public void onInitialize() {
     events();
@@ -33,6 +46,7 @@ public class CobbleNotify implements ModInitializer {
     config.init();
     lang.init();
     Notifications.init();
+    databaseClient = DataBaseFactory.createDataBaseClient(config.getDatabase());
   }
 
 
@@ -48,5 +62,13 @@ public class CobbleNotify implements ModInitializer {
     Events.register();
   }
 
+  public static void runAsync(Runnable task) {
+    CompletableFuture.runAsync(task, EXECUTOR_SERVICE)
+      .orTimeout(30, java.util.concurrent.TimeUnit.SECONDS)
+      .exceptionally(ex -> {
+        ex.printStackTrace();
+        return null;
+      });
+  }
 
 }