+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Added formula, this is more flexible than filters.
 - Added trade notifications, which can be customized in the config.
 - Added global ban option in the config. This will ban pokemons like plushies globally.
+- Added history command to view past actions taken by the mod.ç
+- History supports databases: MongoDB.
 
 ### Bug Fixes